feat: refactored auth, api, added postgres db

+++ server/src/main/java/pl/wymiensie/server/model/Statistics.java
@@ -1,4 +1,4 @@
 package pl.wymiensie.server.model;
 
-public record Statistics(int numberOfUsers, int numberOfBooks) {
+public record Statistics(long numberOfUsers, long numberOfBooks) {
 }
