+++ server/src/main/java/pl/wymiensie/server/service/UserService.java
@@ -23,5 +23,5 @@ public interface UserService {
 
     void deleteUser(UUID id);
 
-    int getNumberOfUsers();
+    long getNumberOfUsers();
 }
